refactor: split directive handlers
